Added code for baselines
1. AnnoMI dataset dir structure
2. Transfer Acc config
3. DER++ baseline
4. DER++ with local attribution replay

+++ networks/my_transformers.py
@@ -794,7 +794,7 @@ class MyBertModel(BertModel):
         head_mask = self.get_head_mask(head_mask, self.config.num_hidden_layers)
 
         embedding_output = self.embeddings(
-            input_ids=input_ids, position_ids=position_ids,  token_type_ids=token_type_ids, inputs_embeds=inputs_embeds
+            input_ids=input_ids.long(), position_ids=position_ids,  token_type_ids=token_type_ids, inputs_embeds=inputs_embeds
         )
 
         encoder_outputs,x_list,h_list = self.compute_encoder_outputs(
